+++ projects/gloo/pkg/syncer/setup/controlplane.go
@@ -3,19 +3,12 @@ package setup
 import (
 	"context"
 	"fmt"
-	"net"
 
 	"github.com/rotisserie/eris"
 	"github.com/solo-io/gloo/pkg/utils"
 	"github.com/solo-io/gloo/pkg/utils/kubeutils"
-	"github.com/solo-io/gloo/projects/gloo/pkg/bootstrap"
-	"github.com/solo-io/gloo/projects/gloo/pkg/xds"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
-	"github.com/solo-io/solo-kit/pkg/api/v1/control-plane/server"
-	xdsserver "github.com/solo-io/solo-kit/pkg/api/v1/control-plane/server"
 	skkube "github.com/solo-io/solo-kit/pkg/api/v1/resources/common/kubernetes"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/reflection"
 )
 
 var (
@@ -23,25 +16,6 @@ var (
 		return eris.Errorf("no port with the name %s found in service %s.%s", portName, svcNamespace, svcName)
 	}
 )
-
-func NewControlPlane(ctx context.Context, grpcServer *grpc.Server, bindAddr net.Addr, kubeControlPlaneCfg bootstrap.KubernetesControlPlaneConfig,
-	callbacks xdsserver.Callbacks, start bool) bootstrap.ControlPlane {
-	snapshotCache := xds.NewAdsSnapshotCache(ctx)
-	xdsServer := server.NewServer(ctx, snapshotCache, callbacks)
-	reflection.Register(grpcServer)
-
-	return bootstrap.ControlPlane{
-		GrpcService: &bootstrap.GrpcService{
-			GrpcServer:      grpcServer,
-			StartGrpcServer: start,
-			BindAddr:        bindAddr,
-			Ctx:             ctx,
-		},
-		SnapshotCache: snapshotCache,
-		XDSServer:     xdsServer,
-		Kube:          kubeControlPlaneCfg,
-	}
-}
 
 // GetControlPlaneXdsPort gets the xDS port from the gloo Service.
 func GetControlPlaneXdsPort(ctx context.Context, svcClient skkube.ServiceClient) (int32, error) {
